maf/requirements.txt
# Microsoft Agent Framework (preview)
agent-framework --pre

# Web server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Bot Framework (MS Teams integration)
botbuilder-core>=4.16.0
botbuilder-integration-aiohttp>=4.16.0
aiohttp>=3.9.0

# Microsoft Graph SDK
msgraph-sdk>=1.12.0
azure-identity>=1.19.0

# Auth
msal>=1.31.0

# Document generation
openpyxl>=3.1.0
python-pptx>=1.0.0

# Configuration
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.28.0
